requirements.txt
numpy
opencv-python
easyocr
pillow